backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
scikit-image==0.22.0
pandas==2.1.3
numpy==1.26.0
rasterio==1.3.9
pyyaml==6.0.1